Add RDL parser, register / bus generator (#8)
A lot of the RDL parsing code and some of the generation code is reused from the Caliptra main repo. It has been moved around a little and cleaned up.

We also changed it to use Tock Registers rather than our own custom type, since we're targeting Tock for our runtime firmware.

This also contains new code is in the xtask/registers.rs file, which contains the pieces to automatically generate a root bus that can be used by the emulator. We can populate the peripherals it uses as we write them (which need to implement the autogenerated traits that are provided).