Run GitHub actions for pull requests; build docs (#11)
Every pull request and the main branch should run the precheckin tests. This is mostly copied from the caliptra-sw repo.

In addition, we also publish docs on merges to main now. An example is published at https://verbose-adventure-9pevejn.pages.github.io/ (these will be moved to a better location when this repo is made public).