If external interrupt is set, keep triggering (#12)
We previously had a workaround to prevent re-triggering an interrupt
while handling an interrupt, but this is unnecessary as the
`mstatus` CSR disables interrupts in the handler until `mret` is
called at the end.

This was causing UART interrupts to not be reasserted even through there
was still data in the buffer to handle.

In addition, we add a loop around handling the UART input buffer (to
avoid needing to wait for another interrupt). We also handle backspaces
correctly now.

We also fix the emulator control address to be in the right place (the
RISC-V code was using a similar address from a Verilator simulated CPU).