Build and execute ROM and Runtime as flat binaries (#21)
Rather than from the ELF files.

We also build the apps and prepend the correct TBF header, and pack that into the binary as well.

This meant we had to copy in the TBF header file from [elf2tab](https://github.com/tock/elf2tab/blob/master/src/header.rs).

We now build the runtime binary twice: once to get the initial size, and then we re-compile / re-link it with the correct offsets so that the apps are loaded as close to the kernel as possible. This helps minimize the RAM requirements.

This fixes a bunch of small bugs related to setting the stacks and capsules correctly in the board file so that the example app will actually execute.

There are some bugs related in the app related to either formatted printing or the timer interface with the capsules that I haven't 100% debugged yet.