Add interrupts support for I3C target (#49)
This maps the internal queues in the I3C emulator peripheral to the
interrupt controller, and modifies the tests so that they no longer
manually poll.

I also added a new package, `romtime`, for drivers and utilities that
are useful for both ROM and Runtime. For now, this contains an
implementation of `print!` and `println!` that output to the emulator
UART immediately, which is useful for both ROM and when running in Tock
kernel when `debug!` may not be available (such as during tests or
interrupts).

We also route IBIs from the TTI register to the I3C controller.

Co-authored-by: Parvathi Bhogaraju <pbhogaraju@microsoft.com>