Image Loading API implementation (#92)
* Image Loading API Implementation

- Implement image loading API interface as described in
https://github.com/chipsalliance/caliptra-mcu-sw/blob/main/docs/src/image_loading.md
- Import checksum API from caliptra-core
- Import mailbox API from caliptra-core only with the commands used for
Image Loading (GetImageOffset, GetImageSize, AuthorizeAndStash)
- Add Fake drivers for DMA, Mailbox and Flash to be used for unit
testing
- Add unit tests for Image Loading and Mailbox
- Fixed bug on syscall driver on DMA and Mailbox. The problem was the
allowed RO or RW buffers shared by the syscall to the driver was not
released after the upcall. It was found that there is need to put the
RO/RW buffers in the generic of share::scope block, otherwise, the
unallow closures associated with RO/RW buffers will not be called after
the call.