rom: Don't allow panics (#108)
In the ROM, we don't want to allow panics: we want to know explicitly
any time we can reach a fatal error.

So we add a similar to test to Caliptra's to scan the ROM ELF for a
symbol we inject when a panic is possible:
https://github.com/chipsalliance/caliptra-sw/blob/main/rom/dev/tests/rom_integration_tests/test_panic_missing.rs

We had to remove the usage of `{:x}` in formatting calls in the ROM,
since current versions of Rust (incorrectly?) think it can panic.

Runtime will be allowed to panic, though we should take care to minimize
the amount of calls, since we cannot reasonably recover from a kernel
panic, but it is probably not possible to remove all panics from Tock.