Fix libapi can not be included as device app dependency (#135)
Problem:
If libapi is included as a dependency of a tock app, the compiler will
throw an exception that libapi is dependent on std lib so it can not be
added.

Solution:
Add no_std in the libapi module lib.rs. Need to remove the
test_checksum_overflow unit test case, since it will cause a stack
overflow when run on a RISC compiler.