Run Caliptra Core by default; autobuild its FW (#139)
Caliptra Core will now always run, since we will start using it soon for
mailbox commands.

To make it easier to use, we now automatically build Caliptra ROM, FW,
and SoC manifest if they are not passed in.

I factored out the `CaliptraBuilder` struct to make it easier for it to
reason about when to automatically build the FW if not provided, when to
build the SoC manifest if not provided, etc.