Add Caliptra mailbox capsule (#145)
This implements the syscall driver we have and connects it with the
Caliptra mailbox, which is connected in the emulator to the Caliptra
emulator.

One problem with the standard Caliptra `SocManager` interface is that it
expects the entire message and response to be buffered by the caller.
When we are passing data through the kernel capsule, we want to avoid
haivng these large buffers.

So, I wrote an adapter layer that uses `impl Iterator<item = u32>`s
instead of requiring a full request or response buffer.

I added a simple test that calls `QUOTE_PCRS`.

I found the source for some of alarms not firing as well: we are not
always calling `alarm.setup()` and `alarm.set_client()` when creating
our `VirtualMuxAlarm`s. I will fix this in a separate PR since we will
need to refactor the way we initialize some pieces so that they use
static initialization and components, since we'll need to use explicit
static lifetimes to call those methods.