Update Recovery Boot Flow Doc (#214)
* Update Recovery Boot Flow Doc

Since MCU can not reset Caliptra core in the event of boot failures,
the Caliptra core will just indicate the failure to the MCU.
The MCU will then switch the active partition to the previous working
one and set the FW_ERROR_FATAL bit. This bit will be an indication to
the platform to reset the full SOC.

The MCU will no longer initiate reset and will leave it up to the
integrator to determine the best time to do so.