I3C and build fixes necessary for FPGA running full recovery mode (#235)
This does some cleanup in the ROM and I3C code to add the necessary initialization for recovery, and adds in some additional parameters (`McuStraps`) that can be passed into the ROM and runtime so the ROM can support more customization (like setting AXI users and watchdog timeouts).

This brings in all of the changes necessary for the FPGA to run the full recovery flow as simply as:

```
cargo xtask fpga-install-kernel-modules && cargo t -p mcu-hw-model --features fpga_realtime -- --test model_fpga_realtime::test::test_new_unbooted --exact --nocapture
```

Apologies for the mess in `hw/model` -- I'm still cleaning that up.