Add xtask for running a ROM on the FPGA without recompiling (#263)
* Add xtask for running a ROM on the FPGA without recompiling

This makes it faster to do tests when changing the ROM, since you can
compile locally, copy to the FPGA, and run without recompiling or
relinking.

I also fixed some of the logic that allows the FPGA ROM to trigger the
FPGA hw model to exit correctly.

This triggered some clippy warnings, which I cleaned up.

* Add Copilot suggested fixes; explicitly run emulator hw model test