Firmware Update (#261)
* Firmware Update

This is the initial implementation for firmware update.

1. Created a new task that spawns a PLDM service that downloads the
firmware update image to a staging area.

The staging area can be implemented by a trait and in this
implementation the image is downloaded to an external SRAM.

2. Currently this patch only supports update of Caliptra and Auth
Manifest.

After the flash image is downloaded, Caliptra is updated using the
Caliptra_fw_load command, while the Auth Manifest is updated using the
set_auth_manifest command

3. Due to the size of the Caliptra image and the auth manifest, it needs
to be sent in chunks through the mailbox. Therefore an API is added to
the mailbox to allow sending mailbox request payload in chunks.

* Addressed feedback

- Move platform specific dma mapping function to platforms/emulator