runtime: Set timers correctly and use mcycle (#290)
The timer counters reset to 0 when the timer reaches the bounds, so they
can't pull double duty as cycle counters as well. Instead, we need to
use `mcycle` and `mcycleh` CSRs to check the current time.

This also corrects the clock for the FPGA and sets up a pattern for
integrators to specify the CPU clock speed.