Implement PLDM Update Agent Retries (#287)
* Implement PLDM Update Agent Retries

When Update Agent sends a packet, it will schedule a timer and wait for
device to respond. If the timer expires, the packet will be resent.

Update agent will retry up to 3 times every 3 seconds. This is done on
both discovery and firmware update state machines.

* Ignore unexpected packet

If the PLDM packet is retried, the device maybe responding to a previous
request. In this case, the Update Agent will be receiving a response
from a previous request. The Update Agent will not process the packet if
it already processed the same packet before, so we ignore the same
packet and continue processing the next packet