Eliminate recursion in flash tests (#327)
* Don't use alarm at all in uart

This was causing the tock$ prompt to keep being printed.

This alarm doesn't seem to be needed at all for us, so I just removed it
altogether.

* Eliminate recursion in flash tests

The flash tests were written with recursive calls to `run()`, which
calls `kernel_loop()`, which calls `alarm()`, which calls operations,
which in turn call `run()`, etc.

This would overflow the stack.

I'm not sure why this wouldn't happen before the prompt fix (also
included); probably something to do with how the alarms were scheduled
if there was another alarm to be processed?