Add emulator SOC requester MCU Mailbox integration test (#341)
* Add emulator SOC requester MCU Mailbox integration tes

Add a transport module that emulates a SOC MCU mailbox requester in the
emulator.

Add integration test "test-mcu-mbox-soc-requester-loopback" where the
emulator would send MCU mailbox messages to the device and device
loopsback the request back to emulator. The emulator would check if the
messages are received and contains the exact data.

Allow MCU Mailbox registers to be shared through several threads through
Arc. This is because the integration test is running on a different
thread with the MCU emulator thread, and they both access the MCU
Mailbox registers

* Fixed missing include