Pass SoC Manifest SVN fuse values to Caliptra (#377)
* Pass SoC Manifest SVN fuse values to Caliptra

- MCU ROM retrieves SoC Manifest SVN and Max SVN values and passes it to
Caliptra using SoC interface.
- Add logic in MCU ROM to monitor Caliptra for Fatal Errors while
waiting for FW to be ready. Caliptra will signal a fatal error if the
SVN check fails. In this way, MCU will propagate the fatal error to SoC.
- Added SoC Manifest SVN to the OTP emulator peripheral
- Add arguments to the emulator runner to pass SoC Manifest SVN
- Add integration test to verify SoC Manifest Verification

* Fix failing FPGA reg_access unit test

Need to specify explicitly pqc_type and also speficy the PK values in
the boot params for fuses