Replace Emulated DMA with AMD DMA driver (#432)
* Replace Emulated DMA with AMD DMA driver

There is a need for an actual DMA implementation that can be supported
in hardware.

For this change, the support for AMD DMA controller is added (see
https://docs.amd.com/r/en-US/pg034-axi-cdma).

Only the simple transfer mode is supported.

The emulated DMA driver code is kept, but the MCU runtime is now using
the AMD DMA controller. The emulator peripheral for AMD DMA is also
created.

* Replace DummaDMA with AxiCDMA in model_emulated

* Remove Emulated DMA code completely