Onboard mctp-capsule-loopback test to FPGA CI (#437)
The `FwId` and `REGISTERED_FW` abstraction doesn't quite work for specifying the different runtime binaries since they are in a different crate for FPGA. There's additional complexity in that a different runtime would also require a different SoC manifest to authorize it. So I added a slightly different mechanism for building and packaging them in the the `all-fw.zip` bundle.

The downside is that the FPGA test is flaky due to it halting when going through the full recovery flow to load MCU runtime.