xtask: Add optional path to `caliptra-sw` (#483)
This adds the optional path to add caliptra-sw to build_test for subsystem. This would be used when building tests with a local reference to caliptra-sw.

This will also delete the test binary archive before building, if it exists. Sometimes the docker build can fail silently, this will force the rsync to fail in those cases.