error-codes: Move error codes to a global module (#515)
This will make it easier to reference and create error codes in FW as well as ROM. It gives a global definition for all fatal error codes. This also follows the pattern used by `caliptra-sw`.

This prefixes all current error codes with `ROM_` because they are currently only used in ROM.