[test] update raw unlock token value and reenable JTAG tests in GHA (#581)
* [caliptra-sw] update caliptra-sw repo version

The latest caliptra-sw repo version contains FPGA stability fixes and a
fix to configure the raw unlock token hash straps. This enables fixing
the raw unlock test.

Signed-off-by: Tim Trippel <ttrippel@google.com>

* fpga: Enable JTAG tests in GHA

Fixes a syntax error that was causing tests to be skipped.

* [test] update raw unlock token value

The raw unlock token was rotated in:
https://github.com/chipsalliance/caliptra-ss/commit/6c2e4e45cc69e0432a961b7d4d4ca187c6d24198

This updates the test to use the new token.

Signed-off-by: Tim Trippel <ttrippel@google.com>

---------

Signed-off-by: Tim Trippel <ttrippel@google.com>
Co-authored-by: Christopher Swenson <chswenson@microsoft.com>